Handlers funcionando. Falta arreglar unos indices que hay querestarles -1 y agregar los comandos para ver estados y configurar el bot.

+++ src/Library/IObservable.cs
@@ -7,7 +7,7 @@ namespace Bankbot
     /// </summary>
     public interface IObservable
     {
-        float Amount { get; set; }
-        float Objective { get; set; }
+        double Amount { get; set; }
+        double Objective { get; set; }
     }
 }
